requirements.txt
requests
torch
transformers
datasets
pandas
scikit-learn
python-box==6.0.2
tqdm
seaborn
ensure==1.0.2
pyYAML
joblib
types-PyYAML
streamlit
newspaper3k
lxml_html_clean
nltk
fastapi[standard]
uvicorn==0.18.3
Jinja2==3.1.2
#numpy<2.0